LSR-805 extract locale parsing and other utility methods to xml parser class
