+++ android/library/src/main/java/com/liferay/mobile/screens/ddl/XSDParser.java
@@ -14,16 +14,11 @@
 
 package com.liferay.mobile.screens.ddl;
 
-import android.os.Parcel;
-
 import com.liferay.mobile.screens.ddl.model.Field;
-import com.liferay.mobile.screens.util.LiferayLocale;
 import com.liferay.mobile.screens.util.LiferayLogger;
 
-import org.w3c.dom.Attr;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
-import org.w3c.dom.NamedNodeMap;
 import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
 import org.xml.sax.InputSource;
@@ -44,7 +39,7 @@ import javax.xml.parsers.ParserConfigurationException;
 /**
  * @author Jose Manuel Navarro
  */
-public class XSDParser implements Parser {
+public class XSDParser extends XMLParser implements FieldParser {
 
 	public List<Field> parse(String xml, Locale locale) throws SAXException {
 		if (xml == null || xml.isEmpty()) {
@@ -65,30 +60,6 @@ public class XSDParser implements Parser {
 		}
 
 		return result;
-	}
-
-	public List<Field> createForm(List<Field> fields, String content) {
-		try {
-			DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();
-			DocumentBuilder builder = factory.newDocumentBuilder();
-			Document document = builder.parse(new InputSource(new StringReader(content)));
-
-			Element root = document.getDocumentElement();
-			return createDocument(root, fields, Locale.ENGLISH);
-		}
-		catch (ParserConfigurationException | IOException | SAXException e) {
-			LiferayLogger.e("Error parsing form", e);
-		}
-		return null;
-	}
-
-	public Field clone(Field field) {
-		Parcel p = Parcel.obtain();
-		p.writeValue(field);
-		p.setDataPosition(0);
-		Field otherField = (Field) p.readValue(Field.class.getClassLoader());
-		p.recycle();
-		return otherField;
 	}
 
 	protected List<Field> processDocument(Document document, Locale locale) {
@@ -218,135 +189,11 @@ public class XSDParser implements Parser {
 	protected Element findMetadataElement(
 		Element dynamicElement, Locale locale, Locale defaultLocale) {
 
-		// Locale matching fallback mechanism: it's designed in such a way to return
-		// the most suitable locale among the available ones. It minimizes the default
-		// locale fallback. It supports input both with one component (language) and
-		// two components (language and country).
-		//
-		// Examples for locale = "es_ES"
-		// 	a1. Matches elements with "es_ES" (full match)
-		//  a2. Matches elements with "es"
-		//  a3. Matches elements for any country: "es_ES", "es_AR"...
-		//  a4. Matches elements for default locale
+		String tagName = "meta-data";
 
-		// Examples for locale = "es"
-		// 	b1. Matches elements with "es" (full match)
-		//  b2. Matches elements for any country: "es_ES", "es_AR"...
-		//  b3. Matches elements for default locale
+		String attrName = "locale";
 
-		NodeList metadataList = dynamicElement.getElementsByTagName("meta-data");
-		int metadataLen = (metadataList == null) ? 0 : metadataList.getLength();
-
-		if (metadataLen == 0) {
-			return null;
-		}
-
-		Element resultElement = null;
-
-		Element metadataElement = getChild(
-			dynamicElement, "meta-data", "locale", locale.toString());
-
-		if (metadataElement != null) {
-			// cases 'a1' and 'b1'
-			resultElement = metadataElement;
-		}
-
-
-		if (resultElement == null) {
-
-			String supportedLocale = LiferayLocale.getSupportedLocaleWithNoDefault(locale.getLanguage());
-			// Pre-final fallback (a2, a3, b2): find any metadata with the portal supported languages
-
-			if (supportedLocale != null) {
-				for (int i = 0; resultElement == null && i < metadataLen; ++i) {
-					Element childElement = (Element) metadataList.item(i);
-					String childLocale = childElement.getAttribute("locale");
-					if (childLocale != null && supportedLocale.equals(childLocale)
-						&& dynamicElement.equals(childElement.getParentNode())) {
-						resultElement = childElement;
-					}
-				}
-			}
-		}
-
-		if (resultElement == null) {
-			// Final fallback (a4, b3): find default metadata
-			resultElement = getChild(
-				dynamicElement, "meta-data", "locale", defaultLocale.toString());
-		}
-
-		return resultElement;
-	}
-
-	protected Map<String, Object> getAttributes(Element element) {
-		NamedNodeMap nodeMap = element.getAttributes();
-		int len = (nodeMap == null) ? 0 : nodeMap.getLength();
-
-		Map<String, Object> result = new HashMap<>(len);
-
-		for (int i = 0; i < len; ++i) {
-			Attr n = (Attr) nodeMap.item(i);
-			result.put(n.getName(), n.getValue());
-		}
-
-		return result;
-	}
-
-	protected Element getChild(Element element, String tagName, String attrName, String attrValue) {
-		List<Element> elements = getChildren(element, tagName, attrName, attrValue);
-		return elements.isEmpty() ? null : elements.get(0);
-	}
-
-	protected List<Element> getChildren(Element element, String tagName, String attrName, String attrValue) {
-		NodeList childList = element.getElementsByTagName(tagName);
-		int len = (childList == null) ? 0 : childList.getLength();
-
-		List<Element> result = new ArrayList<>(len);
-
-		for (int i = 0; i < len; ++i) {
-			Element childElement = (Element) childList.item(i);
-
-			if (attrValue.equals(childElement.getAttribute(attrName)) &&
-				childElement.getParentNode() == element) {
-				result.add(childElement);
-			}
-		}
-
-		return result;
-	}
-
-	private List<Field> createDocument(Element root, List<Field> formFields, Locale locale) {
-		//FIXME use locale
-		List<Field> fields = new ArrayList<>();
-
-
-		NodeList dynamicElementList = root.getElementsByTagName("dynamic-element");
-
-		for (int i = 0; i < dynamicElementList.getLength(); i++) {
-			Element dynamicElement = (Element) dynamicElementList.item(i);
-
-			Field field = getFieldByName(formFields, dynamicElement.getAttribute("name"));
-			if (field != null) {
-				NodeList elementsByTagName = dynamicElement.getElementsByTagName("dynamic-content");
-				Field otherField = field.isRepeatable() ? clone(field) : field;
-				otherField.setCurrentValue(elementsByTagName.item(elementsByTagName.getLength() - 1).getTextContent());
-				if (dynamicElement.getElementsByTagName("dynamic-element").getLength() > 0) {
-					otherField.setFields(createDocument(dynamicElement, field.getFields(), locale));
-				}
-				fields.add(otherField);
-			}
-		}
-
-		return fields;
-	}
-
-	private Field getFieldByName(List<Field> fields, String name) {
-		for (Field field : fields) {
-			if (field.getName().equals(name)) {
-				return field;
-			}
-		}
-		return null;
+		return getLocaleFallback(dynamicElement, locale, defaultLocale, tagName, attrName);
 	}
 
 }
